+++ tests/gossip_required_peer_liveness_chaos.rs
@@ -33,7 +33,7 @@ impl ActorMessageHandlerSync for EchoHandler {
         actor_id: u64,
         type_hash: u32,
         payload: AlignedBytes,
-        correlation_id: Option<u16>,
+        correlation_id: Option<u32>,
     ) -> icanact_remote::Result<Option<ActorResponse>> {
         assert_eq!(actor_id, ACTOR_ID);
         assert_eq!(type_hash, TYPE_HASH);
